refactor: Rss 더미, 타입 분리
