+++ src/types/rss.ts
@@ -1,0 +1,11 @@
+export interface Rss {
+  title: string;
+  link: string;
+  description: string;
+  pubDate: string;
+  author: string;
+  category: string;
+  appbegin: string;
+  appdue: string;
+  budget: number;
+}
